Your commit message

+++ admin/includes/view_all_posts.php
@@ -115,8 +115,8 @@ if(isset($_POST['checkBoxArray'])){
                 echo "<td> $post_comment_count </td>";
                 echo "<td> $post_date </td>";
                 echo "<td><a href='../post.php?p_id={$post_id}'> View Posts </a></td> ";
-                echo "<td><a href='posts.php?delete={$post_id}'> Delete </a></td> ";
-                echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}'> Edit </a></td> ";
+                echo "<td><a onClick=\"javascript: return confirm('Are You sure you want to delete');\" href='posts.php?delete={$post_id}'> Delete </a></td> ";
+                echo "<td><a  href='posts.php?source=edit_post&p_id={$post_id}'> Edit </a></td> ";
 
 
                 echo "</tr>";
